REDME.md
1\. Download the add-on as a ZIP file from Gumroad.

2\. Open Blender and go to Edit > Preferences > Add-ons.

3\. Click the Install button and select the ZIP file you downloaded.

4\. Enable the Blender Render Notification Bot by checking the box next to it in the add-ons list.

**Configuration**

Before you can use the Blender Render Notification Bot, you'll need to set up a Telegram bot and obtain your chat ID. Here's how to do it:

1\. Create a new bot in Telegram by chatting with BotFather. - <https://t.me/BotFather>

2 .Follow the prompts from BotFather to set a name and username for your bot, and to obtain a bot token.

3\. Obtain your chat ID by chatting with userinfobot. - <https://t.me/userinfobot>  (Note down the chat ID, as you'll need it later)  

4\. In Blender, go to Edit > Preferences > Add-ons > Render Notification and enter your bot token and chat ID.

**Usage**

To use the blender_render_notification_bot, simply start a render in Blender. When the render is complete, you'll receive a Telegram message with the render time and other details.

**Support**

If you encounter any issues with the Render Notification add-on, please submit an issue on the GitHub repository.

**License**

The blender_render_notification_bot is released under the MIT license. See LICENSE for more information.